ENGR00139462 Enable ext4 & UBIFS in recovery system.
Gingerbread already support ext4, this patch change
the format command and fix SEGENV when /misc partition
is not avaliable.

UBIFS related script is comment out.

In recovery mode, it default not show the menu, but an image,
if you want to see the menu, please push <Home> key on your device,
or the F1 key in your dell keyboard.

Signed-off-by: Zhang Jiejing <jiejing.zhang@freescale.com>